Abstract Dialog Box out and improved Ui of duke chatbot
