+++ src/main/java/duke/Command.java
@@ -16,6 +16,7 @@ import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.Background;
 import javafx.scene.layout.BackgroundFill;
+import javafx.scene.layout.CornerRadii;
 import javafx.scene.layout.HBox;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Circle;
@@ -239,51 +240,5 @@ public class Command {
 
     private boolean isFindPriority() throws Exception {
         return fullCommand.startsWith("find /p");
-    }
-
-    public static class DialogBox extends HBox {
-        @FXML
-        private Label dialog;
-        @FXML
-        private ImageView displayPicture;
-
-        private DialogBox(String text, Image img) {
-            Circle clip = new Circle(img.getWidth()/3, img.getWidth()/3,50);
-
-            try {
-                FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
-                fxmlLoader.setController(this);
-                fxmlLoader.setRoot(this);
-                fxmlLoader.load();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            dialog.setText(text);
-            displayPicture.setImage(img);
-            displayPicture.setClip(clip);
-            this.setBackground(new Background(new BackgroundFill(Color.AQUA, null, null)));
-        }
-
-        /**
-         * Flips the dialog box such that the ImageView is on the left and text on the right.
-         */
-        private void flip() {
-            ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
-            Collections.reverse(tmp);
-            getChildren().setAll(tmp);
-            setAlignment(Pos.TOP_LEFT);
-            this.setBackground(new Background(new BackgroundFill(Color.DEEPPINK, null, null)));
-        }
-
-        public static DialogBox getUserDialog(String text, Image img) {
-            return new DialogBox(text, img);
-        }
-
-        public static DialogBox getDukeDialog(String text, Image img) {
-            var db = new DialogBox(text, img);
-            db.flip();
-            return db;
-        }
     }
 }
